Swagger now outputs as `YAML` instead of `JSON` (more readable and modifiable).
This change also includes leaves the original references in-tact (so that if the swagger file is modified by hand, the references are there instead of the objects being copied and repeated in the swagger spec).

This also leaves Swagger `$ref`s in place when validating (the swagger-parse was dereferencing the spec and modifying the spec object directly). Using lodash deepClone before running validate to avoid original object modification.

Also replacing `yamljs` with `js-yaml`.

Also fixing missing Swagger `tags`.