Changing the publish trigger
Changing it to run when a release is created, rather than when published (on created is how it was originally in the GitHub Actions example). Now I understand why it might be better that way; I may want to update release notes, without triggering another publish to the NPM and GitHub Packages registries.